+++ tests/test_integration.py
@@ -67,27 +67,6 @@ call_result = [
         "warn": "Getting variable from pressure level data.",
     },
     {
-        # preliminary-back-extension is combined with monthly-means
-        "call": dedent(
-            """\
-            era5cli monthly --variables temperature --startyear 1960 --prelimbe
-            --dryrun"""
-        ),
-        "result": dedent(
-            """\
-            reanalysis-era5-pressure-levels-monthly-means-preliminary-back-extension
-            {'variable': 'temperature', 'year': 1960, 'month': ['01', '02',
-            '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],
-            'time': ['00:00'], 'format': 'netcdf', 'pressure_level': [1, 2, 3,
-            5, 7, 10, 20, 30, 50, 70, 100, 125, 150, 175, 200, 225, 250, 300,
-            350, 400, 450, 500, 550, 600, 650, 700, 750, 775, 800, 825, 850,
-            875, 900, 925, 950, 975, 1000], 'product_type':
-            'reanalysis-monthly-means-of-daily-means'}
-            era5_temperature_1960_monthly.nc"""
-        ),
-        "warn": "The years of the ERA5 preliminary back extension",
-    },
-    {
         # era5-Land is combined with monthly means
         "call": dedent(
             """\